First version of split-and-parse tool
Split is sequential as it must be.

Parse is parallel up to 16 way.

It still produces a deterministic result list.

This can be used to test data parsing the way it is commonly done with
a protocol analyzer breaking the data up (splitter), feeding data to a
Daffodil parser. Except here we use Daffodil to do the splitting also,
just with a simpler splitting schema.

This uses EDIFACT as an "expensive" schema that is worth it to run
in parallel, and the unit tests can be scaled up so you can observe the
speedup from parsing in parallel.